requirements.txt
python-dotenv==0.10.3
python-telegram-bot==11.1.0
dialogflow==0.7.2
vk-api==11.6.1